Create Box_Row.py
practice making graphics with lots of repeated elements. This will help us get familiar with working with multiple shapes and pixel calculations!

Make a line of boxes as shown below, such that the boxes fill the bottom of the canvas. Each box should have a width and height of BOX_SIZE, making a total of 5 boxes perfectly in line with one another: To make each individual box visible instead of making a single rectangular box, add optional arguments to your create rectangle(...) call. The first optional argument is the fill color (in this case "white") and the second optional argument is the outline color (in this case "black").