Create Compute_Average.py
It is often the case that we don't want to hardcode and store large amounts of data directly in the code of our programs. Instead, it is more common for us to transport data using files and read the data from the files to then operate on it. This also lets us operate on different files which are structured the same using the same exact code! The starter code already loads a list of numbers from file. You may find it interesting to read over. 

Your job is to compute the average of the values in number list and to print the result out.