Create The_Game_of_Nimm.py
nstructions
Nimm is an ancient game of strategy that is named after the old German word for "take." It is also called Tiouk Tiouk in West Africa and Tsynshidzi in China. Players alternate taking stones until there are zero left. The game of Nimm goes as follows:

The game starts with a pile of 20 stones between the players

The two players alternate turns

On a given turn, a player may take either 1 or 2 stone from the center pile

The two players continue until the center pile has run out of stones.

The last player to take a stone loses.



Write a program to play Nimm. To make your life easier we have broken the problem down into smaller milestones. You have a lot of time for this program. Take it slowly, piece by piece.